Refactor: remove duplicate interfaces
