+++ src/hooks/usePlatforms.ts
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import platforms from "../data/platforms";
 import apiClient from "../services/api-client";
 import { FetchResponse } from "../services/api-client";
-interface Platform {
+export interface Platform {
   id: number;
   name: string;
   slug: string;
